+++ app/f1/page.tsx
@@ -1,6 +1,7 @@
 import { ConvexHttpClient } from "convex/browser";
 import { api } from "../../convex/_generated/api";
 import { HomeClient } from "../HomeClient";
+import { nextRace } from "../../lib/venues";
 
 export const dynamic = "force-dynamic";
 
@@ -20,12 +21,18 @@ export default async function F1Page({ searchParams }: F1PageProps) {
   }
 
   const convex = new ConvexHttpClient(convexUrl);
-  const initialApprovedSignals = await convex.query(api.actions.approvedVenueCandidates, {});
+  const [initialApprovedSignals, initialScreenings] = await Promise.all([
+    convex.query(api.actions.approvedVenueCandidates, {}),
+    convex.query(api.actions.screeningsForEvent, {
+      eventKey: nextRace.eventKey,
+    }),
+  ]);
 
   return (
     <HomeClient
       initialArea={params.area ?? ""}
       initialApprovedSignals={initialApprovedSignals}
+      initialScreenings={initialScreenings}
       initialInvite={params.invite === "1"}
       basePath="/f1"
     />